Title: 진수 to 진수, Time: 127ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 241224/진수 to 진수/transformation-of-number-system.java
@@ -13,24 +13,24 @@ public class Main {
         int tmp = n;
         int idx = 0;
         while (tmp > 9){
-            ten += (tmp % 10) * Math.pow(a, idx);
+            ten += (tmp % 10) * (int)Math.pow(a, idx);
             idx++;
             tmp /= 10;
         }
-        ten += (tmp % 10) * Math.pow(a, idx);
+        ten += (tmp % 10) * (int)Math.pow(a, idx);
         
         tmp = 0;
         idx = 0;
         // System.out.println(ten);
         while (ten >= b){
-            tmp += (ten % b) * Math.pow(10, idx);
+            tmp += (ten % b) * (int)Math.pow(10, idx);
             idx++;
             // System.out.println(ten % b);
             // tmp *= 10;
             ten /= b;
             // System.out.println(ten);
         }
-        tmp += (ten % b) * Math.pow(10, idx);
+        tmp += (ten % b) * (int)Math.pow(10, idx);
         System.out.print(tmp);
     }
 }
